refactor(#14): 폴더명 변경
