+++ src/pages/Board/BoardDetail/BoardDetail.jsx
@@ -2,23 +2,23 @@ import React, { useRef, useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import { CKEditor } from "@ckeditor/ckeditor5-react";
 import ClassicEditor from "@ckeditor/ckeditor5-build-classic";
-import Slide from "../../components/Common/imgSlide";
-import heart from "../../assets/images/heart.png";
-import commentsubmit from "../../assets/images/commentsubmit.png";
+import Slide from "../../../components/Common/imgSlide";
+import heart from "../../../assets/images/heart.png";
+import commentsubmit from "../../../assets/images/commentsubmit.png";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faEllipsisVertical } from "@fortawesome/free-solid-svg-icons";
-import { fetchcomment, createcomment } from "../../services/api";
-import { fetchPostDetail } from "../../services/postApi";
+import { fetchcomment, createcomment } from "../../../services/api";
+import { fetchPostDetail } from "../../../services/postApi";
 import {
   PostProfileBox,
   ProfileImage,
-} from "../../components/Card/PostCard/PostProfile";
+} from "../../../components/Card/PostCard/PostProfile";
 import {
   Interaction,
   InteractionItem,
-} from "../../components/Common/Interactions";
-import { Wrap } from "../WriteBoard/WriteBoard.style";
-import { deletepost } from "../../services/postApi";
+} from "../../../components/Common/Interactions";
+import { Wrap } from "../BoardWrite/BoardWrite.style";
+import { deletepost } from "../../../services/postApi";
 import {
   CategotyWrap,
   CenteredContainer,
@@ -42,9 +42,9 @@ import {
   CommentInputWrap,
   CommentInput,
   InputImg,
-} from "./DetailPage.style";
+} from "./BoardDetail.style";
 
-function DetailPage() {
+function BoardDetail() {
   const [post, setPost] = useState(null);
   const [commentsData, setCommentData] = useState(null);
   const [commentValue, setCommentValue] = useState("");
@@ -224,4 +224,4 @@ function DetailPage() {
   );
 }
 
-export default DetailPage;
+export default BoardDetail;
