+++ src/pages/Board/BoardDetail/BoardDetail.style.js
